+++ whetstone/compiler/src/main/kotlin/com/freeletics/mad/whetstone/codegen/naventry/NavEntrySubcomponentGenerator.kt
@@ -5,8 +5,10 @@ import com.freeletics.mad.whetstone.codegen.Generator
 import com.freeletics.mad.whetstone.codegen.util.bindsInstanceParameter
 import com.freeletics.mad.whetstone.codegen.util.bundle
 import com.freeletics.mad.whetstone.codegen.util.compositeDisposable
+import com.freeletics.mad.whetstone.codegen.util.contributesToAnnotation
 import com.freeletics.mad.whetstone.codegen.util.coroutineScope
 import com.freeletics.mad.whetstone.codegen.util.internalApiAnnotation
+import com.freeletics.mad.whetstone.codegen.util.internalWhetstoneApi
 import com.freeletics.mad.whetstone.codegen.util.savedStateHandle
 import com.freeletics.mad.whetstone.codegen.util.scopeToAnnotation
 import com.freeletics.mad.whetstone.codegen.util.subcomponentAnnotation
@@ -21,6 +23,10 @@ internal val Generator<NavEntryData>.navEntrySubcomponentFactoryClassName get() 
 
 internal const val navEntrySubcomponentFactoryCreateName = "create"
 
+internal val Generator<NavEntryData>.navEntrySubcomponentFactoryProviderClassName get() = navEntrySubcomponentClassName.nestedClass("ParentComponent")
+
+internal val Generator<NavEntryData>.navEntrySubcomponentFactoryProviderGetterName get() = "get${navEntrySubcomponentClassName.simpleName}"
+
 internal class NavEntrySubcomponentGenerator(
     override val data: NavEntryData,
 ) : Generator<NavEntryData>() {
@@ -29,8 +35,9 @@ internal class NavEntrySubcomponentGenerator(
         return TypeSpec.interfaceBuilder(navEntrySubcomponentClassName)
             .addAnnotation(internalApiAnnotation())
             .addAnnotation(scopeToAnnotation(data.scope))
-            .addAnnotation(subcomponentAnnotation(data.scope))
+            .addAnnotation(subcomponentAnnotation(data.scope, data.parentScope))
             .addType(navEntrySubcomponentFactory())
+            .addType(navEntrySubcomponentFactoryParentComponent())
             .build()
     }
 
@@ -52,6 +59,18 @@ internal class NavEntrySubcomponentGenerator(
         return TypeSpec.interfaceBuilder(navEntrySubcomponentFactoryClassName)
             .addAnnotation(subcomponentFactoryAnnotation())
             .addFunction(createFun)
+            .build()
+    }
+
+    private fun navEntrySubcomponentFactoryParentComponent(): TypeSpec {
+        val getterFun = FunSpec.builder(navEntrySubcomponentFactoryProviderGetterName)
+            .addModifiers(ABSTRACT)
+            .addAnnotation(internalWhetstoneApi)
+            .returns(navEntrySubcomponentFactoryClassName)
+            .build()
+        return TypeSpec.interfaceBuilder(navEntrySubcomponentFactoryProviderClassName)
+            .addAnnotation(contributesToAnnotation(data.parentScope))
+            .addFunction(getterFun)
             .build()
     }
 }